Added local test for get token requset
